+++ src/main/java/com/cburch/logisim/std/io/extra/Slider.java
@@ -11,7 +11,7 @@
  * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
  * for more details.
  *
- * You should have received a copy of the GNU General Public License along 
+ * You should have received a copy of the GNU General Public License along
  * with logisim-evolution. If not, see <http://www.gnu.org/licenses/>.
  *
  * Original code by Carl Burch (http://www.cburch.com), 2011.
@@ -71,7 +71,7 @@ public class Slider extends InstanceFactory {
         else if (sliderPosition > SliderWidth) sliderPosition = SliderWidth;
         if (data.right_to_left) sliderPosition = (byte) (SliderWidth - sliderPosition);
         long value =
-            (long) Math.round(sliderPosition * (Math.pow(2, b.getWidth()) - 1) / SliderWidth);
+            Math.round(sliderPosition * (Math.pow(2, b.getWidth()) - 1) / SliderWidth);
         data.setCurrentValue(value);
         data.setCurrentX(sliderPosition);
         state.getAttributeSet().setValue(ATTR_VALUE, value);
